Move PropertyExtensions from Update to Metadata
